+++ src/qililab/pulse/circuit_to_pulses.py
@@ -1,11 +1,15 @@
 """Class that translates a Qibo Circuit into a PulseSequence"""
+import ast
 from dataclasses import asdict, dataclass
 
 import numpy as np
-from qibo.gates import Gate, M
+from qibo.gates import CZ, Gate, M
 from qibo.models.circuit import Circuit
 
+from qililab import Drag, Park
 from qililab.chip import Chip
+from qililab.chip.nodes import Port, Qubit
+from qililab.config import logger
 from qililab.constants import RUNCARD
 from qililab.pulse.hardware_gates import HardwareGateFactory
 from qililab.pulse.hardware_gates.hardware_gate import HardwareGate
@@ -13,12 +17,10 @@ from qililab.pulse.pulse import Pulse
 from qililab.pulse.pulse_event import PulseEvent
 from qililab.pulse.pulse_schedule import PulseSchedule
 from qililab.settings import RuncardSchema
-from qililab.transpiler import Drag
 from qililab.typings.enums import Line
 from qililab.utils import Factory
 
 
-@dataclass
 class CircuitToPulses:
     """Class that translates a Qibo Circuit into a PulseSequence"""
 
@@ -43,25 +45,82 @@ class CircuitToPulses:
         for circuit in circuits:
             pulse_schedule = PulseSchedule()
             time: dict[int, int] = {}  # restart time
-            readout_gates = circuit.gates_of_type(M)
-            control_gates = [
-                gate for (i, gate) in enumerate(circuit.queue) if i not in [idx for (idx, _) in readout_gates]
-            ]
-            for gate in control_gates:
-                pulse_event, port = self._control_gate_to_pulse_event(time=time, control_gate=gate, chip=chip)
-                if pulse_event is not None:
-                    pulse_schedule.add_event(pulse_event=pulse_event, port=port)
-            for _, readout_gate in readout_gates:
-                for qubit_idx in readout_gate.target_qubits:
-                    readout_pulse_event, port = self._readout_gate_to_pulse_event(
-                        time=time, readout_gate=readout_gate, qubit_idx=qubit_idx, chip=chip
-                    )
-                    if readout_pulse_event is not None:
+
+            for gate in circuit.queue:
+                if isinstance(gate, M):
+                    # handle measurement gates
+                    for qubit_idx in gate.target_qubits:
+                        m_gate = M(qubit_idx)
+                        readout_pulse_event, port = self._readout_gate_to_pulse_event(
+                            time=time, readout_gate=m_gate, qubit_idx=qubit_idx, chip=chip
+                        )
+
                         pulse_schedule.add_event(pulse_event=readout_pulse_event, port=port)
+
+                elif isinstance(gate, CZ):
+                    # CZ sends a SNZ pulse to target in CZ(control, target)
+                    # handle parking and padding for CZ gates
+                    gate = self._parse_check_cz(gate)
+                    if (
+                        chip.get_node_from_qubit_idx(idx=gate.target_qubits[0], readout=False).frequency
+                        < chip.get_node_from_qubit_idx(idx=gate.control_qubits[0], readout=False).frequency
+                    ):
+                        raise ValueError(
+                            f"Attempting to perform {gate.name} on qubits {gate.qubits} by targeting qubit {gate.target_qubits[0]} which has lower frequency than {gate.control_qubits[0]}"
+                        )
+                    parking_gates = self._get_parking_gates(cz=gate, chip=chip)
+                    for parking_gate, _ in parking_gates:
+                        pulse_event, port = self._control_gate_to_pulse_event(
+                            time=time, control_gate=parking_gate, chip=chip
+                        )
+                        pulse_schedule.add_event(pulse_event=pulse_event, port=port)
+                    # add padd time to CZ target qubit to sync it with parking gate
+                    # if there is more than 1 pad time, add max (this is a bit misleading)
+                    pad_time = max((time for _, time in parking_gates), default=0)
+                    if pad_time != 0:
+                        self._update_time(time=time, qubit_idx=gate.target_qubits[0], pulse_time=pad_time)
+
+                # add control gates
+                pulse_event, port = self._control_gate_to_pulse_event(time=time, control_gate=gate, chip=chip)
+                # add pad time at the end of CZ to both target and control
+                # note that we dont need to do this for the control qubit at the beginning of the pulse
+                # since its time is already synced with the target qubit in _control_gate_to_pulse_event
+                if isinstance(gate, CZ) and pad_time != 0:
+                    self._update_time(time=time, qubit_idx=gate.target_qubits[0], pulse_time=pad_time)
+                    self._update_time(time=time, qubit_idx=gate.control_qubits[0], pulse_time=pad_time)
+                pulse_schedule.add_event(pulse_event=pulse_event, port=port)
 
             pulse_schedule_list.append(pulse_schedule)
 
         return pulse_schedule_list
+
+    def _get_park_pad_time(
+        self,
+        park_settings: RuncardSchema.PlatformSettings.GateSettings,
+        cz_settings: RuncardSchema.PlatformSettings.GateSettings,
+    ):
+        """Gets pad time for parking gate
+
+        Args:
+            park_settings (HardwareGate.HardwareGateSettings): settings for the parking gate
+            cz_settings (HardwareGate.HardwareGateSettings): settings for the cz gate for which we need parking
+
+        Returns:
+            pad_time (int): pad time
+
+        Pad time is the extra time for the pulses on the parked qubits before and afer the snz pulse for the
+        CZ gate is applied. Thus if a CZ duration is t_cz and its corresponding parking gate duration is t_p,
+        pad time will be (t_p - t_cz) / 2 - this time will be allocated before and after the CZ's pulse
+        """
+
+        # ideally pad time would be at Park gate definition
+        # TODO find out why mypy complains about the line below
+        # pad_time = (park_settings.duration - 2 * cz_settings.duration + 2 + cz_settings.shape["t_phi"]) / 2
+        t_park = int(park_settings.duration)
+        t_cz = int(cz_settings.duration)
+        t_phi = int(cz_settings.shape["t_phi"])
+        pad_time = (t_park - (2 * t_cz + 2 + t_phi)) / 2
+        return pad_time
 
     def _build_pulse_shape_from_gate_settings(self, gate_settings: HardwareGate.HardwareGateSettings):
         """Build Pulse Shape from Gate settings
@@ -92,9 +151,11 @@ class CircuitToPulses:
         where qubit_pi_amp is the amplitude of the pi pulse for the given qubit calibrated from Rabi.
         The phase will correspond to the rotation around Z from b in R(a,b)
         """
+
         gate_settings = self._get_gate_settings_with_master_values(gate=control_gate)
         pulse_shape = self._build_pulse_shape_from_gate_settings(gate_settings=gate_settings)
-        # TODO: Add CPhase gate
+        # for CZs check if they are possible (defined at runcard) and switch target if needed
+
         qubit_idx = control_gate.target_qubits[0]
         node = chip.get_node_from_qubit_idx(idx=qubit_idx, readout=False)
         port = chip.get_port_from_qubit_idx(idx=qubit_idx, line=Line.DRIVE)
@@ -104,18 +165,55 @@ class CircuitToPulses:
             pulse_time=gate_settings.duration + self.settings.delay_between_pulses,
         )
 
-        # load amplitude, phase for drag pulse from circuit gate parameters
-        if isinstance(control_gate, Drag):
-            amplitude = (control_gate.parameters[0] / np.pi) * gate_settings.amplitude
-            phase = control_gate.parameters[1]
-            # phase is given by b in Drag(a,b) so there should not be any phase defined at gate settings (runcard)
-            if gate_settings.phase is not None:
-                raise ValueError(
-                    "Drag gate should not have setting for phase since the phase depends only on circuit gate parameters"
-                )
+        # get ports and select those we need
+        # get adjacent flux line (drive line) for CZ,Park gates (all others)
+        if isinstance(control_gate, (CZ, Park)):
+            port = chip.get_port_from_qubit_idx(idx=control_gate.target_qubits[0], line=Line.FLUX)
         else:
-            amplitude = float(gate_settings.amplitude)
+            port = chip.get_port_from_qubit_idx(idx=control_gate.target_qubits[0], line=Line.DRIVE)
+
+        # get amplitude from gate settings
+        amplitude = float(gate_settings.amplitude)
+
+        # check that phase is empty for Drag, CZ, Park
+        # also handle specific gate settings
+        if isinstance(control_gate, (Drag, CZ, Park)):
+            # phase should be None
+            phase = gate_settings.phase
+            if gate_settings.phase is not None:
+                raise ValueError(f"{control_gate.name} gate should not have setting for phase")
+
+            # load amplitude, phase for drag pulse from circuit gate parameters
+            if isinstance(control_gate, Drag):
+                amplitude = (control_gate.parameters[0] / np.pi) * amplitude
+                phase = control_gate.parameters[1]
+
+        else:
             phase = float(gate_settings.phase)
+
+        if isinstance(control_gate, CZ):
+            # SNZ duration at gate settings is the SNZ halfpulse duration
+            # should not use pulse duration interchangeably with gate duration
+            cz_duration = 2 * gate_settings.duration + 2 + gate_settings.shape["t_phi"]
+            # sync both qubits start time and set it for the target qubit
+            max_time = max((time[qubit] for qubit in time if qubit in control_gate.qubits), default=0)
+            time[control_gate.target_qubits[0]] = max_time
+
+            # get old time and update time with pulse duration
+            old_time = self._update_time(
+                time=time,
+                qubit_idx=control_gate.target_qubits[0],
+                pulse_time=cz_duration + self.settings.delay_between_pulses,
+            )
+            # sync control qubit time
+            time[control_gate.control_qubits[0]] = time[control_gate.target_qubits[0]]
+
+        else:
+            old_time = self._update_time(
+                time=time,
+                qubit_idx=qubit_idx,
+                pulse_time=gate_settings.duration + self.settings.delay_between_pulses,
+            )
 
         return (
             PulseEvent(
@@ -133,32 +231,60 @@ class CircuitToPulses:
             port,
         )
 
-    def _get_gate_settings_with_master_values(self, gate: Gate):
-        """get gate settings with master values
+    def _get_parking_gates(self, cz: CZ, chip: Chip):
+        """Gets parking gates for CZ's SNZ pulse
 
         Args:
-            gate (Gate): qibo / native gate
+            cz (CZ): CZ gate with qubits (control, target)
+            chip (Chip): chip representation as a graph.
 
         Returns:
-            gate_settings ()
+            list[int]: list of qubit indices to be parked
+
+
+        Qubits to be parked are those that are adjacent to the target qubit, are not the control qubit
+        and have lower frequency than the target.
         """
-        gate_settings = HardwareGateFactory.gate_settings(
-            gate=gate,
-            master_amplitude_gate=self.settings.master_amplitude_gate,
-            master_duration_gate=self.settings.master_duration_gate,
-        )
 
-        # check if duration is an integer value (admit floats with null decimal part)
-        gate_duration = gate_settings.duration
-        if not isinstance(gate_duration, int):  # this handles floats but also settings reading int as np.int64
-            if gate_duration % 1 != 0:  # check decimals
-                raise ValueError(
-                    f"The settings of the gate {gate.name} have a non-integer duration ({gate_duration}ns). The gate duration must be an integer or a float with 0 decimal part"
+        # get parkable qubits
+        target = cz.target_qubits[0]
+        node = chip.get_node_from_qubit_idx(idx=target, readout=False)
+        # get adjacent nodes
+        adj_nodes = chip._get_adjacent_nodes(node)
+        # return adjacent qubits with lower frequency than target not in CZ gate
+        park_qubits = [
+            adj_node.qubit_index
+            for adj_node in adj_nodes
+            if isinstance(adj_node, Qubit)
+            and adj_node.qubit_index not in cz.qubits
+            and adj_node.frequency < node.frequency
+        ]
+
+        park_gates = []
+        # get qubits to park
+        for qubit in park_qubits:
+            park_gate_settings = [
+                settings_gate for settings_gate in self.settings.gates[qubit] if "Park" in settings_gate.name
+            ]
+            if len(park_gate_settings) == 0:
+                logger.warning(
+                    f"Found parking candidate qubit {qubit} for {cz.name} at qubits {cz.qubits} but did not find settings for parking gate at qubit {qubit}"
                 )
-            else:
-                gate_duration = int(gate_duration)
+                continue
+            cz_gate_settings = [
+                settings_gate for settings_gate in self.settings.gates[cz.qubits] if "CZ" in settings_gate.name
+            ][0]
 
-        return gate_settings
+            # get pad time
+            pad_time = self._get_park_pad_time(park_settings=park_gate_settings[0], cz_settings=cz_gate_settings)
+            if pad_time % 1 != 0 or pad_time < 0:
+                raise ValueError(
+                    f"Value pad_time {pad_time} for park gate at {qubit} and CZ {cz.qubits} has nonzero decimal or is negative"
+                )
+
+            park_gates.append((Park(qubit), int(pad_time)))
+
+        return park_gates
 
     def _readout_gate_to_pulse_event(
         self, time: dict[int, int], readout_gate: Gate, qubit_idx: int, chip: Chip
@@ -201,12 +327,58 @@ class CircuitToPulses:
             port,
         )
 
+    def _parse_check_cz(self, cz: CZ):
+        """Checks that given CZ qubits are supported by the hardware (defined in the runcard).
+        Switches CZ(q1,q2) to CZ(q2,q1) if the former is not supported but the later is (note that CZ is symmetric)
+
+        Args:
+            cz (CZ): qibo CZ gate
+
+        Returns:
+            CZ: qibo CZ gate
+        """
+
+        cz_qubits = cz.qubits
+        two_qubit_gates = [qubit for qubit in self.settings.gates.keys() if isinstance(qubit, tuple)]
+        if cz_qubits in two_qubit_gates:
+            return cz
+        elif cz_qubits[::-1] in two_qubit_gates:
+            return CZ(cz_qubits[1], cz_qubits[0])
+        raise NotImplementedError(f"CZ not defined for qubits {cz_qubits}")
+
+    def _get_gate_settings_with_master_values(self, gate: Gate):
+        """get gate settings with master values
+
+        Args:
+            gate (Gate): qibo / native gate
+
+        Returns:
+            gate_settings ()
+        """
+        gate_settings = HardwareGateFactory.gate_settings(
+            gate=gate,
+            master_amplitude_gate=self.settings.master_amplitude_gate,
+            master_duration_gate=self.settings.master_duration_gate,
+        )
+
+        # check if duration is an integer value (admit floats with null decimal part)
+        gate_duration = gate_settings.duration
+        if not isinstance(gate_duration, int):  # this handles floats but also settings reading int as np.int64
+            if gate_duration % 1 != 0:  # check decimals
+                raise ValueError(
+                    f"The settings of the gate {gate.name} have a non-integer duration ({gate_duration}ns). The gate duration must be an integer or a float with 0 decimal part"
+                )
+            else:
+                gate_duration = int(gate_duration)
+
+        return gate_settings
+
     def _update_time(self, time: dict[int, int], qubit_idx: int, pulse_time: int):
         """Create new timeline if not already created and update time.
 
         Args:
             time (Dict[int, int]): Dictionary with the time of each qubit.
-            qubit_idx (int): Index of the qubit.
+            qubit_idx (int | tuple[int,int]): Index of the qubit or index of 2 qubits for 2 qubit gates.
             pulse_time (int): Duration of the puls + wait time.
         """
         if qubit_idx not in time:
@@ -218,12 +390,41 @@ class CircuitToPulses:
         time[qubit_idx] += pulse_time
         return old_time
 
+    def _update_2q_time(self, time: dict[int, int], qubit_idx: tuple[int, int], pulse_time: int):
+        """Create new timeline if not already created and update time.
+
+        Args:
+            time (Dict[int, int]): Dictionary with the time of each qubit.
+            qubit_idx (tuple[int,int]): Index of the 2 qubit gate.
+            pulse_time (int): Duration of the puls + wait time.
+        """
+
+        # get max time if existing, otherwise initialize both times
+        max_time = max((time[qubit] for qubit in time if qubit in qubit_idx), default=0)
+
+        old_time = max_time
+        residue = pulse_time % self.settings.minimum_clock_time
+        if residue != 0:
+            pulse_time += self.settings.minimum_clock_time - residue
+
+        # update time for both qubits / create dict entry if not created
+        time[qubit_idx[0]] = max_time + pulse_time
+        time[qubit_idx[1]] = max_time + pulse_time
+        return old_time
+
     def _instantiate_gates_from_settings(self):
         """Instantiate all gates defined in settings and add them to the factory."""
-        for qubit, gate_settings_list in self.settings.gates.items():
+        for qubits, gate_settings_list in list(self.settings.gates.items()):
+            # parse string tupples for 2 qubit keys
+            if isinstance(qubits, str):
+                qubit_str = qubits
+                qubits = ast.literal_eval(qubit_str)
+                # check for expected output
+                assert isinstance(qubits, tuple) and list(map(type, qubits)) == [int, int]
+                self.settings.gates[qubits] = self.settings.gates.pop(qubit_str)
             for gate_settings in gate_settings_list:
                 settings_dict = asdict(gate_settings)
                 gate_class = HardwareGateFactory.get(name=settings_dict.pop(RUNCARD.NAME))
                 if not gate_class.settings:
                     gate_class.settings = {}
-                gate_class.settings[qubit] = gate_class.HardwareGateSettings(**settings_dict)
+                gate_class.settings[qubits] = gate_class.HardwareGateSettings(**settings_dict)