+++ src/qililab/pulse/hardware_gates/__init__.py
@@ -1,11 +1,11 @@
 """__init__.py"""
+from .cz import CZ
 from .drag import Drag
 from .hardware_gate import HardwareGate
 from .hardware_gate_factory import HardwareGateFactory
 from .i import I
 from .m import M
+from .park import Park
 from .rx import RX
 from .ry import RY
 from .x import X
-from .xy import XY
-from .y import Y